refactor: LOTTO_TICKET_PRICE 상수 LottoTicketsCount로 이동
