+++ src/test/java/lotto/domain/LottoTicketsCountTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static lotto.domain.LottoTicket.LOTTO_TICKET_PRICE;
+import static lotto.domain.LottoTicketsCount.LOTTO_TICKET_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
